refactor: extract capability helpers from AgentWorkspace
